+++ lib/types/stat.dart
@@ -7,6 +7,20 @@ class Stat {
   int timestamp;
   int value;
 
+  static List<Stat>? fromJsonList(List<List>? json) {
+    if (json == null) {
+      return null;
+    }
+    return json.map((e) => Stat.fromJson(e)).toList();
+  }
+
+  static List<dynamic>? toJsonList(List<Stat>? json) {
+    if (json == null) {
+      return null;
+    }
+    return json.map((e) => [e.timestamp, e.value]).toList();
+  }
+
   @override
   String toString() {
     return 'Stat(timestamp: $timestamp, value: $value)';