+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Check formatting in CI (#203)
 * Use Sentry Flutter 6.0.0-beta.1 (#200)
 * Bump Dependencies (#202)