+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Use Sentry Flutter 6.0.0-beta.1 (#200)
 * Bump Dependencies (#202)
 * Fix code generation (#204)
+* In-App Rating (#205)
 
 # 1.0.2