+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 1.0.3
+
 * Check formatting in CI (#203)
 * Use Sentry Flutter 6.0.0-beta.1 (#200)
 * Bump Dependencies (#202)